+++ src/generator.rs
@@ -1,6 +1,6 @@
 //! SCRU128 generator and related items.
 //!
-//! This module is also exported as `scru128::generator` for backward compatibility.
+//! This module is also exported as `scru128::gen` for backward compatibility.
 
 use crate::{Scru128Id, MAX_COUNTER_HI, MAX_COUNTER_LO, MAX_TIMESTAMP};