+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Name of `gen` module back to `generator` to avoid forthcoming `gen` keyword
+  - `gen` remains as an alias to `generator` for backward compatibility
+
 ## v3.0.2 - 2023-09-18
 
 ### Changed